implement simple history
