+++ rnote-engine/src/store/chrono_comp.rs
@@ -90,7 +90,7 @@ impl StrokeStore {
     pub fn keys_sorted_chrono(&self) -> Vec<StrokeKey> {
         let chrono_components = &self.chrono_components;
 
-        let mut keys = self.strokes.keys().collect::<Vec<StrokeKey>>();
+        let mut keys = self.stroke_components.keys().collect::<Vec<StrokeKey>>();
 
         keys.par_sort_unstable_by(|&first, &second| {
             if let (Some(first_chrono), Some(second_chrono)) =
