+++ rnote-engine/src/store/mod.rs
@@ -2,43 +2,32 @@ pub mod chrono_comp;
 pub mod keytree;
 pub mod render_comp;
 pub mod selection_comp;
+pub mod stroke_comp;
 pub mod trash_comp;
-
-use std::collections::VecDeque;
-use std::sync::Arc;
 
 // Re-exports
 pub use chrono_comp::ChronoComponent;
 use keytree::KeyTree;
 pub use render_comp::RenderComponent;
-use rnote_compose::penpath::Segment;
 pub use selection_comp::SelectionComponent;
 pub use trash_comp::TrashComponent;
 
-use crate::pens::penholder::PenStyle;
-use crate::pens::tools::DragProximityTool;
-use crate::strokes::strokebehaviour::GeneratedStrokeImages;
-use crate::strokes::BitmapImage;
-use crate::strokes::Stroke;
-use crate::strokes::StrokeBehaviour;
-use crate::strokes::VectorImage;
-use crate::surfaceflags::SurfaceFlags;
-use crate::{render, Camera};
-use rnote_compose::helpers::{self};
-use rnote_compose::shapes::ShapeBehaviour;
-use rnote_compose::transform::TransformBehaviour;
+use std::collections::VecDeque;
+use std::sync::Arc;
 
-use p2d::bounding_volume::{BoundingSphere, BoundingVolume, AABB};
-use rayon::iter::{ParallelBridge, ParallelIterator};
+use crate::pens::penholder::PenStyle;
+use crate::strokes::strokebehaviour::GeneratedStrokeImages;
+use crate::strokes::Stroke;
+use crate::surfaceflags::SurfaceFlags;
+use crate::Camera;
+use rnote_compose::shapes::ShapeBehaviour;
 use serde::{Deserialize, Serialize};
 use slotmap::{HopSlotMap, SecondaryMap};
-
-use self::render_comp::RenderCompState;
 
 /*
 StrokeStore implements a Entity - Component - System pattern.
 The Entities are the StrokeKey's, which represent a stroke. There are different components for them:
-    * 'strokes': Hold geometric data. These components are special in that they are the primary map. A new stroke must have this component. (could also be called geometric components)
+    * 'stroke_components': Hold geometric data. These components are special in that they are the primary map. A new stroke must have this component. (could also be called geometric components)
     * 'trash_components': Hold state wether the strokes are trashed
     * 'selection_components': Hold state wether the strokes are selected
     * 'chrono_components': Hold state about the time, chronological ordering
@@ -100,8 +89,8 @@ struct HistoryEntry {
 #[serde(default, rename = "stroke_store")]
 pub struct StrokeStore {
     // Components
-    #[serde(rename = "strokes")]
-    strokes: Arc<HopSlotMap<StrokeKey, Arc<Stroke>>>,
+    #[serde(rename = "stroke_components")]
+    stroke_components: Arc<HopSlotMap<StrokeKey, Arc<Stroke>>>,
     #[serde(rename = "trash_components")]
     trash_components: Arc<SecondaryMap<StrokeKey, Arc<TrashComponent>>>,
     #[serde(rename = "selection_components")]
@@ -142,7 +131,7 @@ impl Default for StrokeStore {
         let (tasks_tx, tasks_rx) = futures::channel::mpsc::unbounded::<StoreTask>();
 
         Self {
-            strokes: Arc::new(HopSlotMap::with_key()),
+            stroke_components: Arc::new(HopSlotMap::with_key()),
             trash_components: Arc::new(SecondaryMap::new()),
             selection_components: Arc::new(SecondaryMap::new()),
             chrono_components: Arc::new(SecondaryMap::new()),
@@ -164,7 +153,7 @@ impl Default for StrokeStore {
 
 impl StrokeStore {
     /// The max length of the history
-    pub(crate) const HISTORY_LEN: usize = 100;
+    pub(crate) const HISTORY_MAX_LEN: usize = 100;
 
     pub fn new() -> Self {
         Self::default()
@@ -174,7 +163,7 @@ impl StrokeStore {
     /// needs rendering regeneration after calling
     pub fn import_store(&mut self, store: Self) {
         self.clear();
-        self.strokes = store.strokes;
+        self.stroke_components = store.stroke_components;
         self.trash_components = store.trash_components;
         self.selection_components = store.selection_components;
         self.chrono_components = store.chrono_components;
@@ -187,7 +176,7 @@ impl StrokeStore {
 
     pub fn reload_tree(&mut self) {
         let tree_objects = self
-            .strokes
+            .stroke_components
             .iter()
             .map(|(key, stroke)| (key, stroke.bounds()))
             .collect();
@@ -280,7 +269,7 @@ impl StrokeStore {
     }
 
     fn ptr_eq_history(&self, history_entry: &Arc<HistoryEntry>) -> bool {
-        Arc::ptr_eq(&self.strokes, &history_entry.strokes)
+        Arc::ptr_eq(&self.stroke_components, &history_entry.strokes)
             && Arc::ptr_eq(&self.trash_components, &history_entry.trash_components)
             && Arc::ptr_eq(
                 &self.selection_components,
@@ -291,7 +280,7 @@ impl StrokeStore {
 
     fn history_entry_from_current_state(&self) -> Arc<HistoryEntry> {
         Arc::new(HistoryEntry {
-            strokes: Arc::clone(&self.strokes),
+            strokes: Arc::clone(&self.stroke_components),
             trash_components: Arc::clone(&self.trash_components),
             selection_components: Arc::clone(&self.selection_components),
             chrono_components: Arc::clone(&self.chrono_components),
@@ -299,25 +288,88 @@ impl StrokeStore {
     }
 
     fn import_history_entry(&mut self, history_entry: &Arc<HistoryEntry>) {
-        self.strokes = Arc::clone(&history_entry.strokes);
+        self.stroke_components = Arc::clone(&history_entry.strokes);
         self.trash_components = Arc::clone(&history_entry.trash_components);
         self.selection_components = Arc::clone(&history_entry.selection_components);
         self.chrono_components = Arc::clone(&history_entry.chrono_components);
     }
 
+    /// Saves the current state in the history
     pub fn record(&mut self) {
+        /*
+               log::debug!(
+                   "before record - history len: {}, pos: {:?}",
+                   self.history.len(),
+                   self.history_pos
+               );
+        */
+        self.simple_style_record();
+        /*
+               log::debug!(
+                   "after record - history len: {}, pos: {:?}",
+                   self.history.len(),
+                   self.history_pos
+               );
+        */
+    }
+
+    /// Undos the latest changes
+    pub fn undo(&mut self) {
+        /*
+               log::debug!(
+                   "before undo - history len: {}, pos: {:?}",
+                   self.history.len(),
+                   self.history_pos
+               );
+        */
+        self.simple_style_undo();
+        /*
+               log::debug!(
+                   "after undo - history len: {}, pos: {:?}",
+                   self.history.len(),
+                   self.history_pos
+               );
+        */
+    }
+
+    /// Redos the latest changes. The actual behaviour might differ depending on the history mode (simple style, emacs style, ..)
+    pub fn redo(&mut self) {
+        /*
+               log::debug!(
+                   "before redo - history len: {}, pos: {:?}",
+                   self.history.len(),
+                   self.history_pos
+               );
+        */
+        self.simple_style_redo();
+        /*
+               log::debug!(
+                   "after redo - history len: {}, pos: {:?}",
+                   self.history.len(),
+                   self.history_pos
+               );
+        */
+    }
+
+    fn simple_style_record(&mut self) {
+        // as soon as changes are made, remove the future
+        self.history.truncate(
+            self.history_pos
+                .map(|pos| pos + 1)
+                .unwrap_or(self.history.len()),
+        );
         self.history_pos = None;
 
-        if !self
+        if self
             .history
             .back()
-            .map(|last| self.ptr_eq_history(last))
-            .unwrap_or(false)
+            .map(|last| !self.ptr_eq_history(last))
+            .unwrap_or(true)
         {
             self.history
                 .push_back(self.history_entry_from_current_state());
 
-            if self.history.len() > Self::HISTORY_LEN {
+            if self.history.len() > Self::HISTORY_MAX_LEN {
                 self.history.pop_front();
             }
         } else {
@@ -325,8 +377,73 @@ impl StrokeStore {
         }
     }
 
-    // Needs rendering regeneration after calling
-    pub(crate) fn undo(&mut self) {
+    fn simple_style_undo(&mut self) {
+        let index = match self.history_pos {
+            Some(index) => index,
+            None => {
+                // If we are in the present, we push the current state to the history
+                let current = self.history_entry_from_current_state();
+                self.history.push_back(current);
+
+                self.history.len() - 1
+            }
+        };
+
+        if index > 0 {
+            let prev = Arc::clone(&self.history[index - 1]);
+            self.import_history_entry(&prev);
+
+            self.history_pos = Some(index - 1);
+
+            self.reload_tree();
+        } else {
+            log::debug!("no history, can't undo");
+        }
+    }
+
+    fn simple_style_redo(&mut self) {
+        let index = self.history_pos.unwrap_or(self.history.len() - 1);
+
+        if index < self.history.len() - 1 {
+            let next = Arc::clone(&self.history[index + 1]);
+            self.import_history_entry(&next);
+
+            self.history_pos = Some(index + 1);
+
+            self.reload_tree();
+        } else {
+            log::debug!("no future history entries, can't redo");
+        }
+    }
+
+    /// Saves the current state in the history.
+    /// Only to be used in combination with emacs_style_break_undo_chain() and emacs_style_undo()
+    #[allow(unused)]
+    fn emacs_style_record(&mut self) {
+        self.history_pos = None;
+
+        if self
+            .history
+            .back()
+            .map(|last| !self.ptr_eq_history(last))
+            .unwrap_or(true)
+        {
+            self.history
+                .push_back(self.history_entry_from_current_state());
+
+            if self.history.len() > Self::HISTORY_MAX_LEN {
+                self.history.pop_front();
+            }
+        } else {
+            log::trace!("state has not changed, no need to record");
+        }
+    }
+
+    /// emacs style undo, where the undo operation is pushed to the history as well.
+    /// Needs rendering regeneration after calling
+    /// Only to be used in combination with emacs_style_break_undo_chain() and emacs_style_record()
+    #[allow(unused)]
+    fn emacs_style_undo(&mut self) {
         let index = self.history_pos.unwrap_or(self.history.len());
 
         if index > 0 {
@@ -348,8 +465,11 @@ impl StrokeStore {
         }
     }
 
-    pub(crate) fn break_undo_chain(&mut self) {
-        // move the position to the end, so we can do "undo the undoes", emacs-style
+    /// Breaks the undo chain, to enable redo by undoing the undos, emacs-style.
+    /// Only to be used in combination with emacs_style_undo() and emacs_style_record()
+    #[allow(unused)]
+    fn emacs_style_break_undo_chain(&mut self) {
+        // move the position to the end, so we can do "undo the undos", emacs-style
         self.history_pos = None;
     }
 
@@ -362,7 +482,7 @@ impl StrokeStore {
     pub fn insert_stroke(&mut self, stroke: Stroke) -> StrokeKey {
         let bounds = stroke.bounds();
 
-        let key = Arc::make_mut(&mut self.strokes).insert(Arc::new(stroke));
+        let key = Arc::make_mut(&mut self.stroke_components).insert(Arc::new(stroke));
         self.key_tree.insert_with_key(key, bounds);
         self.chrono_counter += 1;
 
@@ -384,26 +504,14 @@ impl StrokeStore {
         self.render_components.remove(key);
 
         self.key_tree.remove_with_key(key);
-        Arc::make_mut(&mut self.strokes)
+        Arc::make_mut(&mut self.stroke_components)
             .remove(key)
             .map(|stroke| (*stroke).clone())
     }
 
-    /// stroke geometry needs to be updated and rendering regeneration after calling
-    pub fn add_segment_to_brushstroke(&mut self, key: StrokeKey, segment: Segment) {
-        if let Some(Stroke::BrushStroke(brushstroke)) = Arc::make_mut(&mut self.strokes)
-            .get_mut(key)
-            .map(Arc::make_mut)
-        {
-            brushstroke.push_segment(segment);
-
-            self.set_rendering_dirty(key);
-        }
-    }
-
     /// Clears every stroke and every component
     pub fn clear(&mut self) {
-        Arc::make_mut(&mut self.strokes).clear();
+        Arc::make_mut(&mut self.stroke_components).clear();
         Arc::make_mut(&mut self.trash_components).clear();
         Arc::make_mut(&mut self.selection_components).clear();
         Arc::make_mut(&mut self.chrono_components).clear();
@@ -413,585 +521,5 @@ impl StrokeStore {
 
         self.render_components.clear();
         self.key_tree.clear();
-    }
-
-    /// All keys
-    pub fn keys_unordered(&self) -> Vec<StrokeKey> {
-        self.strokes.keys().collect()
-    }
-
-    /// All stroke keys, not including the selection
-    pub fn stroke_keys_unordered(&self) -> Vec<StrokeKey> {
-        self.strokes
-            .keys()
-            .filter_map(|key| {
-                if !(self.trashed(key).unwrap_or(false)) && !(self.selected(key).unwrap_or(false)) {
-                    Some(key)
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Returns the stroke keys in the order that they should be rendered. Does not include the selection.
-    pub fn stroke_keys_as_rendered(&self) -> Vec<StrokeKey> {
-        self.keys_sorted_chrono()
-            .into_iter()
-            .filter_map(|key| {
-                if !(self.trashed(key).unwrap_or(false)) && !(self.selected(key).unwrap_or(false)) {
-                    Some(key)
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<StrokeKey>>()
-    }
-
-    pub fn stroke_keys_as_rendered_intersecting_bounds(&self, bounds: AABB) -> Vec<StrokeKey> {
-        self.keys_sorted_chrono_intersecting_bounds(bounds)
-            .into_iter()
-            .filter(|&key| {
-                !(self.trashed(key).unwrap_or(false)) && !(self.selected(key).unwrap_or(false))
-            })
-            .collect::<Vec<StrokeKey>>()
-    }
-
-    pub fn clone_strokes(&self, keys: &[StrokeKey]) -> Vec<Stroke> {
-        keys.iter()
-            .filter_map(|&key| Some((**self.strokes.get(key)?).clone()))
-            .collect::<Vec<Stroke>>()
-    }
-
-    pub fn insert_vectorimage_bytes_threaded(&mut self, pos: na::Vector2<f64>, bytes: Vec<u8>) {
-        let tasks_tx = self.tasks_tx.clone();
-
-        let all_strokes = self.keys_unordered();
-        self.set_selected_keys(&all_strokes, false);
-
-        self.threadpool.spawn(move || {
-                match String::from_utf8(bytes) {
-                    Ok(svg) => {
-                        match VectorImage::import_from_svg_data(svg.as_str(), pos, None) {
-                            Ok(vectorimage) => {
-                                let vectorimage = Stroke::VectorImage(vectorimage);
-
-                                tasks_tx.unbounded_send(StoreTask::InsertStroke {
-                                    stroke: vectorimage
-                                }).unwrap_or_else(|e| {
-                                    log::error!("tasks_tx.send() failed in insert_vectorimage_bytes_threaded() with Err, {}", e);
-                                });
-                            }
-                            Err(e) => {
-                                log::error!("VectorImage::import_from_svg_data() failed in insert_vectorimage_bytes_threaded() with Err, {}", e);
-                            }
-                        }
-                    }
-                    Err(e) => {
-                        log::error!("from_utf8() failed in thread from insert_vectorimages_bytes_threaded() with Err {}", e);
-                    }
-                }
-            });
-    }
-
-    pub fn insert_bitmapimage_bytes_threaded(&mut self, pos: na::Vector2<f64>, bytes: Vec<u8>) {
-        let tasks_tx = self.tasks_tx.clone();
-
-        let all_strokes = self.keys_unordered();
-        self.set_selected_keys(&all_strokes, false);
-
-        self.threadpool.spawn(move || {
-                match BitmapImage::import_from_image_bytes(&bytes, pos) {
-                    Ok(bitmapimage) => {
-                        let bitmapimage = Stroke::BitmapImage(bitmapimage);
-
-                        tasks_tx.unbounded_send(StoreTask::InsertStroke {
-                            stroke: bitmapimage
-                        }).unwrap_or_else(|e| {
-                            log::error!("tasks_tx.send() failed in insert_bitmapimage_bytes_threaded() with Err, {}", e);
-                        });
-                    }
-                    Err(e) => {
-                        log::error!("BitmapImage::import_from_svg_data() failed in insert_bitmapimage_bytes_threaded() with Err, {}", e);
-                    }
-                }
-            });
-    }
-
-    pub fn insert_pdf_bytes_as_vector_threaded(
-        &mut self,
-        pos: na::Vector2<f64>,
-        page_width: Option<i32>,
-        bytes: Vec<u8>,
-    ) {
-        let tasks_tx = self.tasks_tx.clone();
-
-        let all_strokes = self.keys_unordered();
-        self.set_selected_keys(&all_strokes, false);
-
-        self.threadpool.spawn(move || {
-                match VectorImage::import_from_pdf_bytes(&bytes, pos, page_width) {
-                    Ok(images) => {
-                        for image in images {
-                            tasks_tx.unbounded_send(StoreTask::InsertStroke {
-                                stroke: Stroke::VectorImage(image)
-                            }).unwrap_or_else(|e| {
-                                log::error!("tasks_tx.send() failed in insert_pdf_bytes_as_vector_threaded() with Err, {}", e);
-                            });
-                        }
-                    }
-                    Err(e) => {
-                        log::error!("VectorImage::import_from_pdf_bytes() failed in insert_pdf_bytes_as_vector_threaded() with Err, {}", e);
-                    }
-                }
-            });
-    }
-
-    pub fn insert_pdf_bytes_as_bitmap_threaded(
-        &mut self,
-        pos: na::Vector2<f64>,
-        page_width: Option<i32>,
-        bytes: Vec<u8>,
-    ) {
-        let tasks_tx = self.tasks_tx.clone();
-
-        let all_strokes = self.keys_unordered();
-        self.set_selected_keys(&all_strokes, false);
-
-        self.threadpool.spawn(move || {
-                match BitmapImage::import_from_pdf_bytes(&bytes, pos, page_width) {
-                    Ok(images) => {
-                        for image in images {
-                            let image = Stroke::BitmapImage(image);
-
-                            tasks_tx.unbounded_send(StoreTask::InsertStroke {
-                                stroke: image
-                            }).unwrap_or_else(|e| {
-                                log::error!("tasks_tx.send() failed in insert_pdf_bytes_as_bitmap_threaded() with Err, {}", e);
-                            });
-                        }
-                    }
-                    Err(e) => {
-                        log::error!("BitmapImage::import_from_pdf_bytes() failed in insert_pdf_bytes_as_bitmap_threaded() with Err, {}", e);
-                    }
-                }
-            });
-    }
-
-    /// Needs rendering regeneration after calling
-    pub fn update_geometry_for_stroke(&mut self, key: StrokeKey) {
-        if let Some(stroke) = Arc::make_mut(&mut self.strokes)
-            .get_mut(key)
-            .map(Arc::make_mut)
-        {
-            match stroke {
-                Stroke::BrushStroke(ref mut brushstroke) => {
-                    brushstroke.update_geometry();
-                    self.key_tree.update_with_key(key, stroke.bounds());
-
-                    self.set_rendering_dirty(key);
-                }
-                Stroke::ShapeStroke(_) => {}
-                Stroke::VectorImage(_) => {}
-                Stroke::BitmapImage(_) => {}
-            }
-        }
-    }
-
-    pub fn update_geometry_for_strokes(&mut self, keys: &[StrokeKey]) {
-        keys.iter().for_each(|&key| {
-            self.update_geometry_for_stroke(key);
-        });
-    }
-
-    /// Calculates the width needed to fit all strokes
-    pub fn calc_width(&self) -> f64 {
-        let new_width = if let Some(stroke) = self
-            .strokes
-            .iter()
-            .filter_map(|(key, stroke)| {
-                if let Some(trash_comp) = self.trash_components.get(key) {
-                    if !trash_comp.trashed {
-                        return Some(stroke);
-                    }
-                }
-                None
-            })
-            .max_by_key(|&stroke| stroke.bounds().maxs[0].round() as i32)
-        {
-            // max_by_key() returns the element, so we need to extract the width again
-            stroke.bounds().maxs[0]
-        } else {
-            0.0
-        };
-
-        new_width
-    }
-
-    /// Calculates the height needed to fit all strokes
-    pub fn calc_height(&self) -> f64 {
-        let new_height = if let Some(stroke) = self
-            .stroke_keys_unordered()
-            .into_iter()
-            .filter_map(|key| self.strokes.get(key))
-            .max_by_key(|&stroke| stroke.bounds().maxs[1].round() as i32)
-        {
-            // max_by_key() returns the element, so we need to extract the height again
-            stroke.bounds().maxs[1]
-        } else {
-            0.0
-        };
-
-        new_height
-    }
-
-    /// Generates the enclosing bounds for the given stroke keys
-    pub fn gen_bounds_for_strokes(&self, keys: &[StrokeKey]) -> Option<AABB> {
-        let mut keys_iter = keys.iter();
-        if let Some(&key) = keys_iter.next() {
-            if let Some(first) = self.strokes.get(key) {
-                let mut bounds = first.bounds();
-
-                keys_iter
-                    .filter_map(|&key| self.strokes.get(key))
-                    .for_each(|stroke| {
-                        bounds.merge(&stroke.bounds());
-                    });
-
-                return Some(bounds);
-            }
-        }
-
-        None
-    }
-
-    /// Collects all bounds for the given strokes
-    pub fn bounds_for_strokes(&self, keys: &[StrokeKey]) -> Vec<AABB> {
-        keys.iter()
-            .filter_map(|&key| Some(self.strokes.get(key)?.bounds()))
-            .collect::<Vec<AABB>>()
-    }
-
-    /// Generates a Svg for all strokes as drawn onto the canvas without xml headers or svg roots. Does not include the selection.
-    pub fn gen_svgs_for_strokes(&self, keys: &[StrokeKey]) -> Vec<render::Svg> {
-        keys.iter()
-            .filter_map(|&key| {
-                let stroke = self.strokes.get(key)?;
-
-                match stroke.gen_svg() {
-                    Ok(svgs) => Some(svgs),
-                    Err(e) => {
-                        log::error!(
-                            "stroke.gen_svg() failed in gen_svg_for_strokes() with Err {}",
-                            e
-                        );
-                        None
-                    }
-                }
-            })
-            .collect::<Vec<render::Svg>>()
-    }
-
-    /// Translate the strokes with the offset.
-    /// Rendering needs to be regenerated
-    pub fn translate_strokes(&mut self, strokes: &[StrokeKey], offset: na::Vector2<f64>) {
-        strokes.iter().for_each(|&key| {
-            if let Some(stroke) = Arc::make_mut(&mut self.strokes)
-                .get_mut(key)
-                .map(Arc::make_mut)
-            {
-                {
-                    // translate the stroke geometry
-                    stroke.translate(offset);
-                    self.key_tree.update_with_key(key, stroke.bounds());
-                }
-            }
-        });
-    }
-
-    pub fn translate_strokes_images(&mut self, strokes: &[StrokeKey], offset: na::Vector2<f64>) {
-        strokes.iter().for_each(|&key| {
-            if let Some(render_comp) = self.render_components.get_mut(key) {
-                for image in render_comp.images.iter_mut() {
-                    image.translate(offset);
-                }
-
-                match render::Image::images_to_rendernodes(&render_comp.images) {
-                    Ok(rendernodes) => {
-                        render_comp.rendernodes = rendernodes;
-                    }
-                    Err(e) => log::error!(
-                        "images_to_rendernode() failed in translate_strokes_images() with Err {}",
-                        e
-                    ),
-                }
-            }
-        });
-    }
-
-    /// Rotates the stroke with angle (rad) around the center.
-    /// Rendering needs to be regenerated
-    pub fn rotate_strokes(&mut self, strokes: &[StrokeKey], angle: f64, center: na::Point2<f64>) {
-        strokes.iter().for_each(|&key| {
-            if let Some(stroke) = Arc::make_mut(&mut self.strokes)
-                .get_mut(key)
-                .map(Arc::make_mut)
-            {
-                {
-                    // rotate the stroke geometry
-                    stroke.rotate(angle, center);
-                    self.key_tree.update_with_key(key, stroke.bounds());
-                }
-            }
-        });
-    }
-
-    pub fn rotate_strokes_images(
-        &mut self,
-        strokes: &[StrokeKey],
-        angle: f64,
-        center: na::Point2<f64>,
-    ) {
-        strokes.iter().for_each(|&key| {
-            if let Some(render_comp) = self.render_components.get_mut(key) {
-                render_comp.state = RenderCompState::Dirty;
-
-                for image in render_comp.images.iter_mut() {
-                    image.rotate(angle, center);
-                }
-
-                match render::Image::images_to_rendernodes(&render_comp.images) {
-                    Ok(rendernodes) => {
-                        render_comp.rendernodes = rendernodes;
-                    }
-                    Err(e) => log::error!(
-                        "images_to_rendernode() failed in rotate_strokes() with Err {}",
-                        e
-                    ),
-                }
-            }
-        });
-    }
-
-    /// Scales the strokes with the factor.
-    /// Rendering needs to be regenerated
-    pub fn scale_strokes(&mut self, strokes: &[StrokeKey], scale: na::Vector2<f64>) {
-        strokes.iter().for_each(|&key| {
-            if let Some(stroke) = Arc::make_mut(&mut self.strokes)
-                .get_mut(key)
-                .map(Arc::make_mut)
-            {
-                {
-                    // rotate the stroke geometry
-                    stroke.scale(scale);
-                    self.key_tree.update_with_key(key, stroke.bounds());
-                }
-            }
-        });
-    }
-
-    pub fn scale_strokes_images(&mut self, strokes: &[StrokeKey], scale: na::Vector2<f64>) {
-        strokes.iter().for_each(|&key| {
-            if let Some(render_comp) = self.render_components.get_mut(key) {
-                render_comp.state = RenderCompState::Dirty;
-
-                for image in render_comp.images.iter_mut() {
-                    image.scale(scale);
-                }
-
-                match render::Image::images_to_rendernodes(&render_comp.images) {
-                    Ok(rendernodes) => {
-                        render_comp.rendernodes = rendernodes;
-                    }
-                    Err(e) => log::error!(
-                        "images_to_rendernode() failed in rotate_strokes() with Err {}",
-                        e
-                    ),
-                }
-            }
-        });
-    }
-
-    pub fn scale_strokes_with_pivot(
-        &mut self,
-        strokes: &[StrokeKey],
-        scale: na::Vector2<f64>,
-        pivot: na::Vector2<f64>,
-    ) {
-        self.translate_strokes(strokes, -pivot);
-        self.scale_strokes(strokes, scale);
-        self.translate_strokes(strokes, pivot);
-    }
-
-    pub fn scale_strokes_images_with_pivot(
-        &mut self,
-        strokes: &[StrokeKey],
-        scale: na::Vector2<f64>,
-        pivot: na::Vector2<f64>,
-    ) {
-        self.translate_strokes_images(strokes, -pivot);
-        self.scale_strokes_images(strokes, scale);
-        self.translate_strokes_images(strokes, pivot);
-    }
-
-    /// Resizes the strokes to new bounds.
-    /// Needs rendering regeneration after calling
-    pub fn resize_strokes(&mut self, strokes: &[StrokeKey], new_bounds: AABB) {
-        let old_bounds = match self.gen_bounds_for_strokes(strokes) {
-            Some(old_bounds) => old_bounds,
-            None => return,
-        };
-
-        strokes.iter().for_each(|&key| {
-            if let Some(stroke) = Arc::make_mut(&mut self.strokes)
-                .get_mut(key)
-                .map(Arc::make_mut)
-            {
-                {
-                    // resize the stroke geometry
-                    let old_stroke_bounds = stroke.bounds();
-                    let new_stroke_bounds = helpers::scale_inner_bounds_in_context_new_outer_bounds(
-                        old_stroke_bounds,
-                        old_bounds,
-                        new_bounds,
-                    );
-                    let scale = new_stroke_bounds
-                        .extents()
-                        .component_div(&old_stroke_bounds.extents());
-                    let rel_offset = new_stroke_bounds.center() - old_stroke_bounds.center();
-
-                    // Translate in relation to the outer bounds
-                    stroke.translate(rel_offset - old_stroke_bounds.center().coords);
-                    stroke.scale(scale);
-                    stroke.translate(old_stroke_bounds.center().coords);
-
-                    self.key_tree.update_with_key(key, stroke.bounds());
-                }
-            }
-        });
-    }
-
-    pub fn resize_strokes_images(&mut self, strokes: &[StrokeKey], new_bounds: AABB) {
-        let old_bounds = match self.gen_bounds_for_strokes(strokes) {
-            Some(old_bounds) => old_bounds,
-            None => return,
-        };
-
-        strokes.iter().for_each(|&key| {
-            if let Some(render_comp) = self.render_components.get_mut(key) {
-                render_comp.state = RenderCompState::Dirty;
-
-                for image in render_comp.images.iter_mut() {
-                    // resize the stroke geometry
-                    let old_image_bounds = image.rect.bounds();
-                    let new_image_bounds = helpers::scale_inner_bounds_in_context_new_outer_bounds(
-                        old_image_bounds,
-                        old_bounds,
-                        new_bounds,
-                    );
-                    let scale = new_image_bounds
-                        .extents()
-                        .component_div(&old_image_bounds.extents());
-                    let rel_offset = new_image_bounds.center() - old_image_bounds.center();
-
-                    // Translate in relation to the outer bounds
-                    image.translate(rel_offset - old_image_bounds.center().coords);
-                    image.scale(scale);
-                    image.translate(old_image_bounds.center().coords);
-                }
-
-                match render::Image::images_to_rendernodes(&render_comp.images) {
-                    Ok(rendernodes) => {
-                        render_comp.rendernodes = rendernodes;
-                    }
-                    Err(e) => log::error!(
-                        "images_to_rendernode() failed in resize_strokes() with Err {}",
-                        e
-                    ),
-                }
-            }
-        });
-    }
-
-    /// Returns all keys below the y_pos
-    pub fn keys_below_y_pos(&self, y_pos: f64) -> Vec<StrokeKey> {
-        self.strokes
-            .iter()
-            .filter_map(|(key, stroke)| {
-                if stroke.bounds().mins[1] > y_pos {
-                    Some(key)
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<StrokeKey>>()
-    }
-
-    /// Needs rendering regeneration for current viewport after calling
-    pub fn drag_strokes_proximity(&mut self, drag_proximity_tool: &DragProximityTool) {
-        let sphere = BoundingSphere {
-            center: na::Point2::from(drag_proximity_tool.pos),
-            radius: drag_proximity_tool.radius,
-        };
-
-        #[allow(dead_code)]
-        fn calc_distance_ratio(
-            pos: na::Vector2<f64>,
-            tool_pos: na::Vector2<f64>,
-            radius: f64,
-        ) -> f64 {
-            // Zero when right at drag_proximity_tool position, One when right at the radius
-            (1.0 - (pos - tool_pos).magnitude() / radius).clamp(0.0, 1.0)
-        }
-
-        Arc::make_mut(&mut self.strokes)
-            .iter_mut()
-            .par_bridge()
-            .filter_map(|(key, stroke)| -> Option<StrokeKey> {
-                match Arc::make_mut(stroke) {
-                    Stroke::BrushStroke(brushstroke) => {
-                        if sphere.intersects(&brushstroke.path.bounds().bounding_sphere()) {
-                            for segment in brushstroke.path.iter_mut() {
-                                let segment_sphere = segment.bounds().bounding_sphere();
-
-                                if sphere.intersects(&segment_sphere) {
-                                    /*                                     match segment {
-                                        Segment::QuadBez { start, cp, end } => {
-                                            start.pos += drag_proximity_tool.offset
-                                                * calc_distance_ratio(
-                                                    start.pos,
-                                                    drag_proximity_tool.pos,
-                                                    drag_proximity_tool.radius,
-                                                );
-                                            *cp += drag_proximity_tool.offset
-                                                * calc_distance_ratio(
-                                                    *cp,
-                                                    drag_proximity_tool.pos,
-                                                    drag_proximity_tool.radius,
-                                                );
-                                            end.pos += drag_proximity_tool.offset
-                                                * calc_distance_ratio(
-                                                    end.pos,
-                                                    drag_proximity_tool.pos,
-                                                    drag_proximity_tool.radius,
-                                                );
-                                            return Some(key);
-                                        }
-                                    } */
-                                    return Some(key);
-                                }
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-
-                None
-            })
-            .collect::<Vec<StrokeKey>>()
-            .iter()
-            .for_each(|&key| {
-                self.update_geometry_for_stroke(key);
-            });
     }
 }
